feat: implement BM25 search for Japanese documents with hybrid search as default
This commit introduces BM25 (Best Matching 25) keyword-based search capabilities
to complement the existing vector search, with special support for Japanese text
tokenization using MeCab/fugashi.

Key changes:
- Add BM25 indexing with inverted index, vocabulary, and document statistics
- Implement Japanese tokenizer using fugashi with fallback to simple tokenization
- Make hybrid search (combining vector + BM25) the default search mode
- Move Japanese tokenizer dependencies (fugashi, unidic-lite, jaconv) to standard dependencies
- Move RAG benchmark dependencies (datasets, scipy, scikit-learn, pandas) to dev dependencies
- Add comprehensive tests for BM25 functionality

The hybrid search approach provides better results by combining:
- Vector search: captures semantic similarity
- BM25 search: handles exact keyword matches
Results are normalized and weighted (default: 70% vector, 30% BM25)

Technical details:
- BM25 parameters: k1=1.2, b=0.75 (standard values)
- Japanese tokenization includes POS filtering and stop word removal
- Database schema extended with BM25-specific tables
- Foreign key constraints properly handled in clear operations

This enhancement significantly improves search accuracy for Japanese documents
while maintaining backward compatibility with existing vector-only search.

+++ src/oboyu/indexer/bm25_indexer.py
@@ -233,7 +233,7 @@ class BM25Indexer:
             "document_count": self.document_count,
             "vocabulary_size": self.get_vocabulary_size(),
             "total_terms": sum(self.collection_frequencies.values()),
-            "avg_document_length": self.total_document_length / max(self.document_count, 1),
+            "avg_document_length": int(self.total_document_length / max(self.document_count, 1)),
         }
     
     def _get_term_positions(self, text: str, term: str) -> List[int]:
